Added error-handling and comments
Added a while loop that incapsulates all code in main to re-promt user with login screen if an invalid login was attemped. Instead of just quitting the program.

Also added an else after admin pin check to achieve the same thing by not quitting the program if the wrong pin was entered.

NOTE! Added hardcoded username "user" before UserMenu gets called so that the UserMenu doesn't get called no matter wht is entered. Altough we need to change this so we check if user exists in database and that he entered pin matches that users pin before callin UserMenu().

Added some clarifying comments as well to increase readability