Added checks and mechanics to freeze and unfreeze users.
-Added variable to save the unfreeze time to

-Added some clarifying comments to separate login, admin and user parts of the code

-Before any login attempt is made we check if the user is frozen or not.
If it is frozen it prints out error message and checks if user can be unfrozen and unfreezes them if yes.
NOTE! This could be optimized in the future to perhaps do the "can be unfrozen check" before so user don't get an error even when their "time out" time has passed as we do the check in the bottom of the code.

-Changed in the loop so if user enters pin wrong 3 times, account is frozen