Added option 6 - to change pin code
Users can now change pin code after a successful login, just like in a real bank/atm.
Method asks for current pin and checks if correct then proceeds to ask for a new pin, followed by a confirmation check to see that you've entered it correctly.
Lastly it then calls the DbHelper method to save the pin to the database. With error handling built in