Added methods to promt user to press the Enter key before proceeding
The two methods PressEnter() are the same with the slight difference that if you send in a string when calling the method it'll print that.
Did this solution since we want to press enter to go to the menu in several places in the program but in some places we just want to press enter to continue.
Now we can choose if and what to promt the user with each time