Re structured my OpenNewAccount method.
Summary:
Changed some wording, some logic. Mainly re-structured what happened when in the code and fixed some error-handling. To now account for wrongful inputs. And cleaned up the code - made it more concise and easier to read.

Detailed:
Added method to check if an account already exists with the same name that user is trying to create to avoid duplicates.

Renamed vacation account to foreign account

Changed so that SEK always is set to the default currency. That way selected currency is declared outside of the foreign currency (old "vacation") so we don't need to set currency to SEK twice.

Added error handling on when user selects foreign currency with a simple while(true) loop and a break when input is correct.

I changed when accounts was created to the bottom part of the method to avoid duplicate declarations of currency. And to better readability.

Now asks user if they want to make an initial deposit regardless of if its a foreign currency account or not. With added error-handling with the same while(true) loop.

My old code to promt user to press enter had disapeared and been replaced with a  Console.Readline(). So I made a method in MenuFunctions that only accepts the Enter key as an input. We can now reuse it in all out code. More info on that in my MenuFunctions commit.