Added error handling to the login screen
Wrapped the main login screen in a while loop that doesn't break out until name and pin isn't null. To avoid crash when just pressing enter.

Before any login code is run we check if the user exists at all with DoesUserExist(). This is to avoid crashes when entering a name that isn't in the database.

Removed code that for some reason was commented away (UnFreezeUser).

Changed some formatting in the printed text to be more coherent.

Added clarifying comments to old and new code.