Added a method copied to handle option navigation like the menu navigation in the UserMenu()
Created OptionsNavigation() method that takes in an array of strings as options that you can cycle between and choose one of the said options.
Will return an int depending on which options you chose so if you have 10 options (10 things in the array) and pressed [Enter] on option 4 it returns 4. So you can use an if or a switch to check which option was chosen.
Takes in a phrase that will be written at the top of the code.