Added the new admin methods to the main menu.
Now upon launch we check if an admin exists in the database at all. If there are none we assume its the first time setup and as such we let the user create a new admin.

This is saved in the new Admin table so it'll only appear on the first launch on a new computer (or when connected to a new db).