Added Admin methods.
IsAdminCreated() simply checks if there is any user in the admin table and returns true or false.

VerifyAdminLogin() verifies so the admin login matches the name and pin in the database

CreateAdmin() lets the user create a new admin user with their chosen name and pin. Error handling is built in so you have to comply with a 4-digit pin and enter the same pin twice in a row for it to accept it. To ensure no mispresses.