added some fixing to all the console.writelines so it follows the same pattern thru out the program.
Also added some headers and footers from MenuFunctions to make it look better.

Added "Main Menu" to the top of main menu so the user know which menu is main menu.

Also fixed so you can use the arrowkeys when chosing an option in all methods and menus.

Used the method OptionNavigation in all methods so it is easier to understand the code. Main Menu still has its own arrowkey method though.