Minor fixes
-Minor fixes
-Removed unused using commands at the top of the code.
-Added Thread.Sleep to some places to make code readable before disapearing

-Logout
Changed the logout option so it now goes back to the login screen instead of exiting the program.
The while loop in UserMenu() now runs on isLoggedIn instead of just true. (Set to true from the start). When log out is chosen it turns false, ending the menu loop.
So HandleMenuSelection now returns a bool.

-TransferMoney
Added Math.Round() to TransferMoney() where balance is printed to round printing down to 2 decimals (no rounding done on any calculations) just to make it look nicer.
Also just changed the Transfer Successful message to look more coherent.

-ChangePin
-Changed so style is according to the rest of the program. Header prints out and clears window on each loop - even when faulty input has been made.
-Clarified error messages.
-Added logic so pin must be 4 digits and now checks so the consecutive pins aren't null. (Previously you could set pin to nothing and couldn't login anymore).