Changed method to array
Changed so instead of having a method to keep all the menu options, we just have an array instead. This shortens down the code quite a bit so it's easier to read.

Had to change some logic then as the method was 1-indexed and arrays are 0-indexed. But all is working the same way only that it's counting from 0-6 instead of 1-7.

Made HandleMenuSelection's access modifier private as it's only used in this class.